Improve upload and replace file performance
First of all the image dimension extractor is removed, as it is obsolete.
The RecordMonitor is triggered every time the index entry is updated,
which is enough to track changes in the file and update the image
sizes accordingly.

Additionally we avoid downloading the file from S3 when we had the
file locally already when it was uploaded by adding them to the
temporary file cache that already exists.

We also use a new and better TYPO3 API to calculate the image size,
which relies on a PHP function instead of an imagemagick call.