Add sanitize function for file name
It turned out, that without the sanitize of a file name some issues occur.
For example a media asset created from a youtube video, which uses the
title of the video as file name and includes a colon, will cause some
problems. One is, that it could not be deleted.

This will affect also folder name, thus the sanitize function is added.