[BUGFIX] Always clean temporary files if created by getFileForLocalProcessing
It has turned out, that files always get "downloaded" to fal-tempfile-{hash}
into directory typo3temp/var/transient, but not cleaned accordingly.

If the method AmazonS3Driver::getFileForLocalProcessing gets called with
the second parameter to false, the file will not be added to the array
$temporaryFiles and thus not deleted on __destruct. As the a file always
gets "downloaded" via the copy statement, the condition to check if
writable is misused and with this, the folder typo3temp/var/transient
gets flooded by temporary files.